Revert "Making player funds into floats"
This reverts commit a4505609ed0264d6e477fe4e0148ec28eda59cbc.

+++ Agromica/Assets/Scripts/Player.cs
@@ -8,8 +8,8 @@ using UnityEngine.UI;
 /// </summary>
 public class Player : MonoBehaviour
 {
-    public float currentMoney;
-    public float currentDebt;
+    public int currentMoney;
+    public int currentDebt;
     public Text inventoryDisplay;
     public Dictionary<string, int> cropInventory = new Dictionary<string, int>();
